Airflow/requirements.txt
kafka-python
#   
pandas
# pyspark==3.5.0
apache-airflow
# apache-airflow-providers-apache-spark
elasticsearch